feat: react memo를 사용해 타이핑 중일 때 contentEidtalbe이 적용된 div의 리렌더링 막기
리액트의 vDOM과 충돌을 막기위해 리렌더링을 막아주었습니다.
