+++ src/app/(main)/note/[id]/_components/content/line-info.tsx
